+++ test/unit/Core.mocks.js
@@ -22,7 +22,7 @@ class CoreMock {
       init: this._sandbox.stub().resolves(),
       start: this._sandbox.stub().resolves(),
       stop: this._sandbox.stub().resolves(),
-      restart: this._sandbox.stub().resolves(),
+      restartServer: this._sandbox.stub().resolves(),
       settings: {
         get: this._sandbox.stub(),
         set: this._sandbox.stub()
@@ -35,14 +35,16 @@ class CoreMock {
         warn: this._sandbox.stub(),
         error: this._sandbox.stub()
       },
-      onChangeSettings: this._sandbox.stub(),
-      onLoadMocks: this._sandbox.stub(),
-      addCustomRouter: this._sandbox.stub(),
-      addCustomSetting: this._sandbox.stub(),
+      onChangeSettings: this._sandbox.stub().returns(() => {}),
+      onChangeMocks: this._sandbox.stub().returns(() => {}),
+      addRouter: this._sandbox.stub(),
+      addSetting: this._sandbox.stub(),
       behaviors: {
+        ids: [],
         currentFromCollection: "foo-current",
         collection: "foo-behaviors-collection"
       },
+      fixtures: {},
       serverError: null,
       _eventEmitter: {
         on: this._sandbox.stub(),